feat: Allow to enabled/disable support Ingersses individually.
This splits the value "ingress.enabled" into the following two values:

- "ingress.favicon.enabled"
- "ingress.minio.enabled"

This way the Ingress resources can be individually enabled or disabled as needed
by the operator.

Migration:

If the value "ingress.enabled" has been changed or explicitly configured, then
this value has now to be applied to the two new values. The value
"ingress.enabled" has to be removed.