feat: Upgrade portal-consumer / portal-server to version 0.50.0
This change extends the portal consumer, so that it is possible to configure a
base URL for assets. This shall support the usage of an S3 compatible storage
which is publishing via an external domain.

New configuration options in the sub-chart:

- "nubusPortalConsumer.assetsBaseUrl"

The value should only be configured if one wants to refer to portal assets via
an external domain.