feat: Remove portal related defaults regarding the log level configuration
The reason is that the respective components charts have been updated so that
they already come with the correct defaults out of the box, so that the umbrella
chart does not have to overwrite these defaults.

There is no migration needed.

Users can still adjust the log levels in the same way as before.