feat: Update portal charts to version 0.51.0
This does add the support for Feature Toggles for the portal-server and the
portal-frontend.

The feature toggles can be configured via the following configuration
parameters:

```
nubusPortalServer:
  portalServer:
    featureToggles:
      notifications_api: false
```

The configuration is also provided via the API to the portal frontend, so that
the toggles can influence the feature related implementation in both components.

The main purpose is to support in-development features or preliminary features
to be rolled out early on.