feat: Update portal subcharts to version 0.55.0
The main feature from this upgrade is the new User API endpoint for the portal
server.

It will enable future frontend features by providing more details about the
current user. The new endpoint is reachable as `/univention/portal/api/v1/me`.

This also adds configuration of the type "UdmClient" into the subchart
`portal-server`. The configuration is already based on the target structure for
secrets and client configuration, see "docs/values-client.md" in this
repository. Other subcharts will be moved towards this structure in the future.

The authentication configuration for the UDM Rest API access is also using the
new target for the handling of authentication configuration.