fix: Update portal components to version 0.66.1
This change does include a fix in the portal frontend for the feature "Quick
Links". A missing image for an entry in a drop down was rendered as a
questionmark. This has been replaced by a transparent image to ensure proper
aligenment of the menu entries without the placeholder image.

This change does not need any special steps in upgrade and migration scenarios.

Issue: univention/dev/internal/team-nubus#1161