ci: add upgrade-tests as a CI Job
The upgrade tests install an old version of Nubus. This is usually the
previous minor version release. The current version is taken from the
`SEMANTIC_VERSION`. All already released versions are collected from the
git tags.

Before doing the upgrade to the new version the CI will run all tests
tagged with pre_upgrade in our e2e pipeline.

This right now includes a single test which sets up 2FA for a user.
After the upgrade in addition to the normal e2e tests all tests
tagged with post_upgrade are run, which currently includes a test
that logs in with the user configured in the pre_upgrade test.

Issue: univention/dev/internal/team-nubus#1535